Removed unused parts of the OAuth2 authentification
Started to implement Oauth2 but get stuck :(. I am currently not sure if
Oauth2 is really needed. For now I skip the Oauth2 protocol to exchange the
authentification information and stay with a simple JWT approach.
The user can now register clients (as in Oauth2) and than login with the
client and will directly get the JWT which later used for authentification.
I will look into this later if Oauth2 is really needed.